requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
certifi==2023.5.7
charset-normalizer==3.1.0
Django==4.2.1
django-extensions==3.2.1
djangorestframework==3.14.0
idna==3.4
pytz==2023.3
razorpay==1.3.0
requests==2.31.0
sqlparse==0.4.4
typing-extensions==4.6.2
urllib3==2.0.2
